fix a few warnings

+++ sources/opl3/adl/player.cc
@@ -90,7 +90,7 @@ std::vector<std::string> Player::enumerate_emulators()
 
     std::vector<std::string> names(32);
     size_t count = 0;
-    for (size_t i = 0, n = names.size(); i < n; ++i) {
+    for (unsigned i = 0, n = (unsigned)names.size(); i < n; ++i) {
         if (adl_switchEmulator(pl.get(), i) == 0) {
             names[i] = adl_chipEmulatorName(pl.get());
             count = i + 1;
